+++ crates/fuel-core/src/executor.rs
@@ -2,7 +2,10 @@
 #[allow(clippy::cast_possible_truncation)]
 #[cfg(test)]
 mod tests {
-    use crate::database::Database;
+    use crate::database::{
+        Database,
+        RelayerReadDatabase,
+    };
     use fuel_core_executor::{
         executor::{
             block_component::PartialBlockComponent,
@@ -20,6 +23,7 @@ mod tests {
             ContractsRawCode,
             Messages,
         },
+        transactional::AtomicView,
         StorageAsMut,
     };
     use fuel_core_types::{
@@ -119,13 +123,20 @@ mod tests {
         Rng,
         SeedableRng,
     };
-    use std::ops::DerefMut;
+    use std::{
+        ops::DerefMut,
+        sync::Arc,
+    };
 
     fn create_executor(
         database: Database,
         config: Config,
-    ) -> Executor<Database, Database> {
-        Executor::test(database, config)
+    ) -> Executor<Database, RelayerReadDatabase> {
+        Executor {
+            database_view_provider: database.clone(),
+            relayer_view_provider: RelayerReadDatabase::new(database),
+            config: Arc::new(config),
+        }
     }
 
     pub(crate) fn setup_executable_script() -> (Create, Script) {
@@ -312,6 +323,7 @@ mod tests {
 
     mod coinbase {
         use super::*;
+        use fuel_core_storage::transactional::AtomicView;
 
         #[test]
         fn executor_commits_transactions_with_non_zero_coinbase_generation() {
@@ -412,7 +424,8 @@ mod tests {
             }
 
             let (asset_id, amount) = producer
-                .database
+                .database_view_provider
+                .latest_view()
                 .contract_balances(recipient, None, None)
                 .next()
                 .unwrap()
@@ -491,7 +504,8 @@ mod tests {
                 panic!("Invalid coinbase transaction");
             }
             let (asset_id, amount) = producer
-                .database
+                .database_view_provider
+                .latest_view()
                 .contract_balances(recipient, None, None)
                 .next()
                 .unwrap()
@@ -524,16 +538,11 @@ mod tests {
             let producer = create_executor(Default::default(), config);
 
             let result = producer
-                .execute_without_commit(
-                    ExecutionTypes::DryRun(Components {
-                        header_to_produce: Default::default(),
-                        transactions_source: OnceTransactionsSource::new(vec![
-                            script.into()
-                        ]),
-                        gas_limit: u64::MAX,
-                    }),
-                    Default::default(),
-                )
+                .execute_without_commit(ExecutionTypes::DryRun(Components {
+                    header_to_produce: Default::default(),
+                    transactions_source: OnceTransactionsSource::new(vec![script.into()]),
+                    gas_limit: u64::MAX,
+                }))
                 .unwrap();
             let ExecutionResult { block, .. } = result.into_result();
 
@@ -609,7 +618,8 @@ mod tests {
                 .unwrap();
             assert_eq!(validated_block.transactions(), produced_txs);
             let (asset_id, amount) = validator
-                .database
+                .database_view_provider
+                .latest_view()
                 .contract_balances(recipient, None, None)
                 .next()
                 .unwrap()
@@ -883,14 +893,11 @@ mod tests {
             transactions: vec![tx.clone()],
         };
 
-        let mut block_db_transaction = producer.database.transaction();
-
         let ExecutionData {
             skipped_transactions,
             ..
         } = producer
             .execute_block(
-                &mut block_db_transaction,
                 ExecutionType::Production(PartialBlockComponent::from_partial_block(
                     &mut block,
                 )),
@@ -908,10 +915,8 @@ mod tests {
         ));
 
         // Produced block is valid
-        let mut block_db_transaction = verifier.database.transaction();
         verifier
             .execute_block(
-                &mut block_db_transaction,
                 ExecutionType::Validation(PartialBlockComponent::from_partial_block(
                     &mut block,
                 )),
@@ -921,9 +926,7 @@ mod tests {
 
         // Invalidate the block with Insufficient tx
         block.transactions.insert(block.transactions.len() - 1, tx);
-        let mut block_db_transaction = verifier.database.transaction();
         let verify_result = verifier.execute_block(
-            &mut block_db_transaction,
             ExecutionType::Validation(PartialBlockComponent::from_partial_block(
                 &mut block,
             )),
@@ -953,13 +956,11 @@ mod tests {
             ],
         };
 
-        let mut block_db_transaction = producer.database.transaction();
         let ExecutionData {
             skipped_transactions,
             ..
         } = producer
             .execute_block(
-                &mut block_db_transaction,
                 ExecutionType::Production(PartialBlockComponent::from_partial_block(
                     &mut block,
                 )),
@@ -973,10 +974,8 @@ mod tests {
         ));
 
         // Produced block is valid
-        let mut block_db_transaction = verifier.database.transaction();
         verifier
             .execute_block(
-                &mut block_db_transaction,
                 ExecutionType::Validation(PartialBlockComponent::from_partial_block(
                     &mut block,
                 )),
@@ -988,9 +987,7 @@ mod tests {
         block
             .transactions
             .insert(block.transactions.len() - 1, Transaction::default_test_tx());
-        let mut block_db_transaction = verifier.database.transaction();
         let verify_result = verifier.execute_block(
-            &mut block_db_transaction,
             ExecutionType::Validation(PartialBlockComponent::from_partial_block(
                 &mut block,
             )),
@@ -1041,13 +1038,11 @@ mod tests {
             transactions: vec![tx.clone()],
         };
 
-        let mut block_db_transaction = producer.database.transaction();
         let ExecutionData {
             skipped_transactions,
             ..
         } = producer
             .execute_block(
-                &mut block_db_transaction,
                 ExecutionType::Production(PartialBlockComponent::from_partial_block(
                     &mut block,
                 )),
@@ -1065,10 +1060,8 @@ mod tests {
         ));
 
         // Produced block is valid
-        let mut block_db_transaction = verifier.database.transaction();
         verifier
             .execute_block(
-                &mut block_db_transaction,
                 ExecutionType::Validation(PartialBlockComponent::from_partial_block(
                     &mut block,
                 )),
@@ -1080,9 +1073,7 @@ mod tests {
 
         // Invalidate block by adding transaction with not existing coin
         block.transactions.insert(block.transactions.len() - 1, tx);
-        let mut block_db_transaction = verifier.database.transaction();
         let verify_result = verifier.execute_block(
-            &mut block_db_transaction,
             ExecutionType::Validation(PartialBlockComponent::from_partial_block(
                 &mut block,
             )),
@@ -2264,7 +2255,7 @@ mod tests {
     }
 
     /// Helper to build database and executor for some of the message tests
-    fn make_executor(messages: &[&Message]) -> Executor<Database, Database> {
+    fn make_executor(messages: &[&Message]) -> Executor<Database, RelayerReadDatabase> {
         let mut database = Database::default();
         let database_ref = &mut database;
 
@@ -2333,24 +2324,22 @@ mod tests {
         let message_data = message_from_input(&tx.inputs()[1], 0);
         let messages = vec![&message_coin, &message_data];
 
-        let mut block = PartialFuelBlock {
+        let block = PartialFuelBlock {
             header: Default::default(),
             transactions: vec![tx.into()],
         };
 
         let exec = make_executor(&messages);
-        let mut block_db_transaction = exec.database.transaction();
-        assert_eq!(block_db_transaction.all_messages(None, None).count(), 2);
+        let view = exec.database_view_provider.latest_view();
+        assert!(!view.message_is_spent(&message_coin.nonce).unwrap());
+        assert!(!view.message_is_spent(&message_data.nonce).unwrap());
 
-        let ExecutionData {
+        let ExecutionResult {
             skipped_transactions,
             ..
         } = exec
-            .execute_block(
-                &mut block_db_transaction,
-                ExecutionType::Production(PartialBlockComponent::from_partial_block(
-                    &mut block,
-                )),
+            .execute_and_commit(
+                ExecutionBlock::Production(block),
                 ExecutionOptions {
                     utxo_validation: true,
                 },
@@ -2359,18 +2348,11 @@ mod tests {
         assert_eq!(skipped_transactions.len(), 0);
 
         // Successful execution consumes `message_coin` and `message_data`.
-        assert_eq!(block_db_transaction.all_messages(None, None).count(), 0);
-        assert!(block_db_transaction
-            .message_is_spent(&message_coin.nonce)
-            .unwrap());
-        assert!(block_db_transaction
-            .message_is_spent(&message_data.nonce)
-            .unwrap());
+        let view = exec.database_view_provider.latest_view();
+        assert!(view.message_is_spent(&message_coin.nonce).unwrap());
+        assert!(view.message_is_spent(&message_data.nonce).unwrap());
         assert_eq!(
-            *block_db_transaction
-                .coin(&UtxoId::new(tx_id, 0))
-                .unwrap()
-                .amount(),
+            *view.coin(&UtxoId::new(tx_id, 0)).unwrap().amount(),
             amount + amount
         );
     }
@@ -2396,24 +2378,22 @@ mod tests {
         let message_data = message_from_input(&tx.inputs()[1], 0);
         let messages = vec![&message_coin, &message_data];
 
-        let mut block = PartialFuelBlock {
+        let block = PartialFuelBlock {
             header: Default::default(),
             transactions: vec![tx.into()],
         };
 
         let exec = make_executor(&messages);
-        let mut block_db_transaction = exec.database.transaction();
-        assert_eq!(block_db_transaction.all_messages(None, None).count(), 2);
+        let view = exec.database_view_provider.latest_view();
+        assert!(!view.message_is_spent(&message_coin.nonce).unwrap());
+        assert!(!view.message_is_spent(&message_data.nonce).unwrap());
 
-        let ExecutionData {
+        let ExecutionResult {
             skipped_transactions,
             ..
         } = exec
-            .execute_block(
-                &mut block_db_transaction,
-                ExecutionType::Production(PartialBlockComponent::from_partial_block(
-                    &mut block,
-                )),
+            .execute_and_commit(
+                ExecutionBlock::Production(block),
                 ExecutionOptions {
                     utxo_validation: true,
                 },
@@ -2422,20 +2402,10 @@ mod tests {
         assert_eq!(skipped_transactions.len(), 0);
 
         // We should spend only `message_coin`. The `message_data` should be unspent.
-        assert_eq!(block_db_transaction.all_messages(None, None).count(), 1);
-        assert!(block_db_transaction
-            .message_is_spent(&message_coin.nonce)
-            .unwrap());
-        assert!(!block_db_transaction
-            .message_is_spent(&message_data.nonce)
-            .unwrap());
-        assert_eq!(
-            *block_db_transaction
-                .coin(&UtxoId::new(tx_id, 0))
-                .unwrap()
-                .amount(),
-            amount
-        );
+        let view = exec.database_view_provider.latest_view();
+        assert!(view.message_is_spent(&message_coin.nonce).unwrap());
+        assert!(!view.message_is_spent(&message_data.nonce).unwrap());
+        assert_eq!(*view.coin(&UtxoId::new(tx_id, 0)).unwrap().amount(), amount);
     }
 
     #[test]
@@ -2599,13 +2569,11 @@ mod tests {
         };
 
         let exec = make_executor(&[&message]);
-        let mut block_db_transaction = exec.database.transaction();
         let ExecutionData {
             skipped_transactions,
             ..
         } = exec
             .execute_block(
-                &mut block_db_transaction,
                 ExecutionType::Production(PartialBlockComponent::from_partial_block(
                     &mut block,
                 )),
@@ -2626,9 +2594,7 @@ mod tests {
 
         // Produced block is valid
         let exec = make_executor(&[&message]);
-        let mut block_db_transaction = exec.database.transaction();
         exec.execute_block(
-            &mut block_db_transaction,
             ExecutionType::Validation(PartialBlockComponent::from_partial_block(
                 &mut block,
             )),
@@ -2641,9 +2607,7 @@ mod tests {
         // Invalidate block by return back `tx2` transaction skipped during production.
         block.transactions.insert(block.transactions.len() - 1, tx2);
         let exec = make_executor(&[&message]);
-        let mut block_db_transaction = exec.database.transaction();
         let res = exec.execute_block(
-            &mut block_db_transaction,
             ExecutionType::Validation(PartialBlockComponent::from_partial_block(
                 &mut block,
             )),